Subscribe to AgreementFulfilled in execute_pending_agreements
